src/main/java/eu/jev/springmvcrest/services/CategoryServiceImpl.java
package eu.jev.springmvcrest.services;

import eu.jev.springmvcrest.api.v1.mapper.CategoryMapper;
import eu.jev.springmvcrest.api.v1.model.CategoryDTO;
import eu.jev.springmvcrest.repositories.CategoryRepository;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class CategoryServiceImpl implements CategoryService {

    private final CategoryRepository categoryRepository;

    private final CategoryMapper categoryMapper;

    public CategoryServiceImpl(CategoryRepository categoryRepository, CategoryMapper categoryMapper) {
        this.categoryRepository = categoryRepository;
        this.categoryMapper = categoryMapper;
    }

    @Override
    public List<CategoryDTO> getAllCategories() {
        return categoryMapper.categoryListToCategoryDTOList(categoryRepository.findAll());
    }

    @Override
    public CategoryDTO getCategoryByName(String name) {
        return categoryMapper.categoryToCategoryDTO(categoryRepository.findByName(name));
    }
}
